gcab: hide the --space option, it's not necessary anymore
Listen the interweb!!! you don't need to reserve space to add
a signature anymore.

I repeat again, you _don't_ need to reserve header space to add a
signature anymore with a somewhat recent signtool.exe. Even if some
old msdn documentation still says so, and plenty of pages.

Although the PKCS#7 data is appended at the end of the cabinet, they
still use the reserved data field, and move the rest of the content,
but nevermind, it's just IO after all, right...

Oh, and osslsigncode doesn't need it either.