configure.ac: don't use disable-fast-install
We're not sure why this is here, and it causes automake to install a
libtool wrapper script named gcab into /usr/bin instead of gcab itself,
which is pretty broken. So rather than debug and figure out what's
wrong, I will just remove it and lo! the problem goes away.