libgcab/gcab-file.c: Don't encode timezone in generated files.
Whilst working on the Reproducible Builds effort [0], we noticed
that gcab generates .cab files that are not reproducible. This
affects packages such as fwupd.

This is because the embedded dates are encoded with the current
timezone.

localtime is expressed relative to the user's specified timezone,
whilst gmtime is in UTC. This is important so that users building the
exact same contents in different timezones will end up with the same
exact .cab file.

  [0] https://reproducible-builds.org/

Signed-off-by: Chris Lamb <lamby@debian.org>

https://bugzilla.gnome.org/show_bug.cgi?id=786435