Add gcab_file_new_with_bytes()
At the moment fwupd decompresses each archive into a folder in /tmp, then reads
back the files into memory so that it can parse them. This is fragile, and
somewhat insecure if not done carefully.

This new API allows us to do two new things:

* Create an archive without a GFile to read from
* Decompress an archive to memory, without writing any files on disk

Allowing GCab to decompress to a GBytes buffer means we can make the fwupd code
a lot simpler and safer.